+++ output/changelog/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- Generated at: 2025-12-07T11:54:13.178431
-- Total time: 215.40 seconds
+- Generated at: 2026-01-01T23:39:38.519517
+- Total time: 54.17 seconds
 - Features: 7
 - Models: 20